+++ Frends.LDAP.SearchObjects/Frends.LDAP.SearchObjects/Definitions/Enums.cs
@@ -61,3 +61,33 @@ public enum LDAPVersion
     /// </summary>
     V3,
 }
+
+/// <summary>
+/// Enumeration of file encoding options.
+/// </summary>
+public enum ContentEncoding
+{
+    // Pragma for self-explanatory enum attributes.
+#pragma warning disable 1591, SA1602
+    UTF8,
+    Default,
+    ASCII,
+    WINDOWS1252,
+#pragma warning restore 1591, SA1602
+    /// <summary>
+    /// Other enables users to add other encoding options as string.
+    /// </summary>
+    Other,
+}
+
+/// <summary>
+/// Enumeration of return types.
+/// </summary>
+public enum ReturnType
+{
+// Pragma for self-explanatory enum attributes.
+#pragma warning disable 1591, SA1602
+    String,
+    Byte,
+#pragma warning restore 1591, SA1602
+}