+++ Frends.LDAP.SearchObjects/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.2.0] - 2025-03-17
+### Added
+- Added support for custom attributes in the output. You can now define which attributes you want to get from the LDAP search.
+- Added support for getting attributes in byte array format.
+- Added encoding options for string handling for attribute values.
+
 ## [2.1.0] - 2025-01-02
 ### Fixed
 - Fixed issue with AttributeSet not having all values that were returned from LDAP search.